+++ core-macos/lib/nlohmann/json.hpp
