+++ core-macos/src/auth/CMakeLists.txt
@@ -1,0 +1,3 @@
+file(GLOB SOURCES *.cpp)
+
+add_library(auth ${SOURCES})
