+++ core-macos/src/cloud/CMakeLists.txt
@@ -1,0 +1,4 @@
+file(GLOB SOURCES *.cpp)
+
+include_directories(../../lib)
+add_library(cloud ${SOURCES})
