+++ core-macos/src/core/CMakeLists.txt
@@ -1,0 +1,6 @@
+include_directories(include)
+
+file(GLOB SOURCES *.cpp)
+
+include_directories(../../lib)
+add_library(core ${SOURCES})
