+++ core-macos/src/core/filesystem.cpp
@@ -4,16 +4,14 @@
 #include <dirent.h>
 #include <sys/types.h>
 #include <sys/stat.h>
-#include "../../../lib/json/json.hpp"
-#include "../../settings.h"
+#include "include/filesystem.h"
+#include "nlohmann/json.hpp"
+#include "../settings.h"
 
 using namespace std;
 using json = nlohmann::json;
 
 namespace filesystem {
-
-
-
     string createDirectory(string jso) {
         json input;
         json output;
