+++ core-macos/src/core/log.cpp
@@ -1,0 +1,3 @@
+#include "include/log.h"
+
+std::mutex log::_mutex;
