+++ core-macos/src/core/storage.cpp
@@ -22,8 +22,9 @@
 
 #include <iostream>
 #include <fstream>
-#include "../../../lib/json/json.hpp"
-#include "../../settings.h"
+#include "include/storage.h"
+#include "nlohmann/json.hpp"
+#include "../settings.h"
 
 using namespace std;
 using json = nlohmann::json;
