+++ core-macos/src/ping/CMakeLists.txt
@@ -1,0 +1,3 @@
+file(GLOB SOURCES *.cpp)
+
+add_library(ping ${SOURCES})
