+++ core-macos/src/platform/include/fd.h
@@ -1,0 +1,13 @@
+#ifndef FD_H
+#define FD_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+    void fd_opendlg(char** outPath);
+    void fd_savedlg(char** outPath);
+#ifdef __cplusplus
+}
+#endif
+
+#endif
